+++ cmd/engram/bootstrap.go
@@ -550,7 +550,11 @@ layer instead, for example:
 
 %s
 
-Do not skip this step.`, engramVersion(), cmd, layer, layerExample, skillManagementGuidance)
+%s
+
+%s
+
+Do not skip this step.`, engramVersion(), cmd, layer, layerExample, memoryTierGuidance, memoryWriteSafetyGuidance, skillManagementGuidance)
 }
 
 func runBootstrapGemini(cmd *cobra.Command, _ []string) error {